finishing_rp

+++ app/Http/Controllers/PasswordResetController.php
@@ -14,7 +14,7 @@ class PasswordResetController extends Controller
 {
     public function showResetForm()
     {
-        return view('auth.passwords.email');  // form buat masukin email
+        return view('auth.password.forget-password');  // form buat masukin email abis klik forget password
     }
 
     public function sendResetLinkEmail(Request $request)
@@ -28,9 +28,11 @@ class PasswordResetController extends Controller
         }
 
         // sementara aku bikin tokennya random dulu
-        $token = Str::random(60);
+        $token = rand(100000, 999999);
+        // $user->token = $token;
+        // $user->save();
 
-        //tokennya jd dimaukin tbel forget pass
+        //tokennya jd dimasukin tabel forget pass
         DB::table('password_resets')->insert([
             'email' => $request->email,
             'token' => $token,
@@ -38,41 +40,53 @@ class PasswordResetController extends Controller
         ]);
 
         //ngirim email forget pass
-        Mail::to($user->email)->send(new ResetPasswordMail($token));
+        // Mail::to($user->email)->send(new ResetPasswordMail($token));
+
+        Mail::send('auth.password.email', ['token' => $token], function ($message) use ($user) {
+            $message->to($user->email);
+            $message->subject('Reset Password');
+        });
 
         return back()->with('message', 'Link forget password telah dikirimkan ke email anda.');
     }
 
     public function showResetPasswordForm($token)
     {
-        return view('auth.passwords.reset', ['token' => $token]);  //form reset password
+        return view('auth.password.reset-password', ['token' => $token]);  //form reset password abis isi token, pass baru
     }
 
     public function resetPassword(Request $request)
     {
+        // Validasi input dari user
         $request->validate([
             'email' => 'required|email',
             'password' => 'required|confirmed|min:6',
             'token' => 'required'
         ]);
 
-        //sesuaiin token di db dulu
+        // Cek token dan email di database
         $passwordReset = DB::table('password_resets')->where([
             ['token', $request->token],
             ['email', $request->email],
         ])->first();
 
+        // Jika token atau email tidak valid
         if (!$passwordReset) {
             return back()->withErrors(['email' => 'Token atau email anda tidak valid.']);
         }
 
-        // pass di update
+        // Ambil pengguna berdasarkan email
         $user = User::where('email', $request->email)->first();
-        $user->update(['password' => bcrypt($request->password)]);
 
-        // klo sukses, token dihapus
-        DB::table('password_resets')->where('email', $request->email)->delete();
+        if ($user) {
+            $user->update(['password' => $request->password]);
 
-        return redirect('/login')->with('message', 'Password anda berhasil diubah!');
+            DB::table('password_resets')->where('email', $request->email)->delete();
+
+            return redirect('/loginasesi')->with('message', 'Password anda berhasil diubah!');
+        }
+
+        return back()->withErrors(['email' => 'Pengguna tidak ditemukan.']);
     }
+
 }
